backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
requests==2.32.3
websockets==13.1
pandas==2.2.3
numpy==2.2.1
ta==0.11.0
